third_party/blink/web_tests/flag-specific/highdpi/virtual/scroll-unification/fast/forms/select-popup/popup-menu-resize-after-open-expected.txt

FAIL popupWindow.innerWidth should be 100. Was 150.
FAIL popupWindow.innerWidth should be 150. Was 225.
PASS successfullyParsed is true

TEST COMPLETE

o1
o2
